BeamFileSystem implementation instead of IOChannelFactory
