+++ sdks/python/apache_beam/io/fileio_test.py
@@ -16,7 +16,7 @@
 # limitations under the License.
 #
 
-"""Unit tests for local and GCS sources and sinks."""
+"""Unit tests for file sinks."""
 
 import glob
 import logging
@@ -26,58 +26,14 @@ import tempfile
 import unittest
 
 import hamcrest as hc
-import mock
 
 import apache_beam as beam
 from apache_beam import coders
 from apache_beam.io import fileio
+from apache_beam.io.filesystem import CompressedFile
 from apache_beam.test_pipeline import TestPipeline
 from apache_beam.transforms.display import DisplayData
 from apache_beam.transforms.display_test import DisplayDataItemMatcher
-
-
-class TestChannelFactory(unittest.TestCase):
-
-  @mock.patch('apache_beam.io.fileio.gcsio')
-  def test_size_of_files_in_glob_complete(self, *unused_args):
-    # Prepare mocks.
-    gcsio_mock = mock.MagicMock()
-    fileio.gcsio.GcsIO = lambda: gcsio_mock
-    file_names = ['gs://bucket/file1', 'gs://bucket/file2']
-    gcsio_mock.size_of_files_in_glob.return_value = {
-        'gs://bucket/file1': 1,
-        'gs://bucket/file2': 2
-    }
-    expected_results = {
-        'gs://bucket/file1': 1,
-        'gs://bucket/file2': 2
-    }
-    self.assertEqual(
-        fileio.ChannelFactory.size_of_files_in_glob(
-            'gs://bucket/*', file_names),
-        expected_results)
-    gcsio_mock.size_of_files_in_glob.assert_called_once_with('gs://bucket/*')
-
-  @mock.patch('apache_beam.io.fileio.gcsio')
-  def test_size_of_files_in_glob_incomplete(self, *unused_args):
-    # Prepare mocks.
-    gcsio_mock = mock.MagicMock()
-    fileio.gcsio.GcsIO = lambda: gcsio_mock
-    file_names = ['gs://bucket/file1', 'gs://bucket/file2']
-    gcsio_mock.size_of_files_in_glob.return_value = {
-        'gs://bucket/file1': 1
-    }
-    gcsio_mock.size.return_value = 2
-    expected_results = {
-        'gs://bucket/file1': 1,
-        'gs://bucket/file2': 2
-    }
-    self.assertEqual(
-        fileio.ChannelFactory.size_of_files_in_glob(
-            'gs://bucket/*', file_names),
-        expected_results)
-    gcsio_mock.size_of_files_in_glob.assert_called_once_with('gs://bucket/*')
-    gcsio_mock.size.assert_called_once_with('gs://bucket/file2')
 
 
 # TODO: Refactor code so all io tests are using same library
@@ -115,16 +71,16 @@ class _TestCaseWithTempDirCleanUp(unittest.TestCase):
 class TestCompressedFile(_TestCaseWithTempDirCleanUp):
 
   def test_seekable(self):
-    readable = fileio._CompressedFile(open(self._create_temp_file(), 'r'))
+    readable = CompressedFile(open(self._create_temp_file(), 'r'))
     self.assertFalse(readable.seekable)
 
-    writeable = fileio._CompressedFile(open(self._create_temp_file(), 'w'))
+    writeable = CompressedFile(open(self._create_temp_file(), 'w'))
     self.assertFalse(writeable.seekable)
 
   def test_tell(self):
     lines = ['line%d\n' % i for i in range(10)]
     tmpfile = self._create_temp_file()
-    writeable = fileio._CompressedFile(open(tmpfile, 'w'))
+    writeable = CompressedFile(open(tmpfile, 'w'))
     current_offset = 0
     for line in lines:
       writeable.write(line)
@@ -132,7 +88,7 @@ class TestCompressedFile(_TestCaseWithTempDirCleanUp):
       self.assertEqual(current_offset, writeable.tell())
 
     writeable.close()
-    readable = fileio._CompressedFile(open(tmpfile))
+    readable = CompressedFile(open(tmpfile))
     current_offset = 0
     while True:
       line = readable.readline()
@@ -299,52 +255,6 @@ class TestFileSink(_TestCaseWithTempDirCleanUp):
     os.remove(res2)
     with self.assertRaises(Exception):
       list(sink.finalize_write(init_token, [res1, res2]))
-
-  @mock.patch('apache_beam.io.fileio.ChannelFactory.rename')
-  @mock.patch('apache_beam.io.fileio.gcsio')
-  def test_rename_batch(self, *unused_args):
-    # Prepare mocks.
-    gcsio_mock = mock.MagicMock()
-    fileio.gcsio.GcsIO = lambda: gcsio_mock
-    fileio.ChannelFactory.rename = mock.MagicMock()
-    to_rename = [
-        ('gs://bucket/from1', 'gs://bucket/to1'),
-        ('gs://bucket/from2', 'gs://bucket/to2'),
-        ('/local/from1', '/local/to1'),
-        ('gs://bucket/from3', 'gs://bucket/to3'),
-        ('/local/from2', '/local/to2'),
-    ]
-    gcsio_mock.copy_batch.side_effect = [[
-        ('gs://bucket/from1', 'gs://bucket/to1', None),
-        ('gs://bucket/from2', 'gs://bucket/to2', None),
-        ('gs://bucket/from3', 'gs://bucket/to3', None),
-    ]]
-    gcsio_mock.delete_batch.side_effect = [[
-        ('gs://bucket/from1', None),
-        ('gs://bucket/from2', None),
-        ('gs://bucket/from3', None),
-    ]]
-
-    # Issue batch rename.
-    fileio.ChannelFactory.rename_batch(to_rename)
-
-    # Verify mocks.
-    expected_local_rename_calls = [
-        mock.call('/local/from1', '/local/to1'),
-        mock.call('/local/from2', '/local/to2'),
-    ]
-    self.assertEqual(fileio.ChannelFactory.rename.call_args_list,
-                     expected_local_rename_calls)
-    gcsio_mock.copy_batch.assert_called_once_with([
-        ('gs://bucket/from1', 'gs://bucket/to1'),
-        ('gs://bucket/from2', 'gs://bucket/to2'),
-        ('gs://bucket/from3', 'gs://bucket/to3'),
-    ])
-    gcsio_mock.delete_batch.assert_called_once_with([
-        'gs://bucket/from1',
-        'gs://bucket/from2',
-        'gs://bucket/from3',
-    ])
 
 
 if __name__ == '__main__':
