+++ sdks/python/apache_beam/io/iobase.py
@@ -591,8 +591,8 @@ class Sink(HasDisplayData):
   single record from the bundle and ``close()`` which is called once
   at the end of writing a bundle.
 
-  See also ``beam.io.fileio.FileSink`` which provides a simpler API for writing
-  sinks that produce files.
+  See also ``apache_beam.io.fileio.FileSink`` which provides a simpler API
+  for writing sinks that produce files.
 
   **Execution of the Write transform**
